fix(billing-types): SMI-5119 remove unpublishable @skillsmith/billing-types package
The W5-W8 consolidated publish failed because @skillsmith/billing-types@0.1.0
could not publish — npm OIDC trusted-publishing requires a package to already
exist on npm, a chicken-and-egg for brand-new packages (existing packages were
token-bootstrapped before SMI-4540 deleted all tokens). Cascade: mcp-server@0.5.3
+ enterprise@0.1.2 were skipped, and cli@0.6.3 shipped referencing a
non-existent mcp-server@^0.5.3.

Root cause: billing-types was consumed ONLY via `import type` (erased at compile
time, zero runtime require) yet declared as a runtime dependency. A types-only
package consumed via import type never needs to be published.

Resolution (user-approved no-token refactor):
- Delete packages/billing-types/ entirely
- mcp-server: inline the StripeWebhookHandler/StripeWebhookResult interfaces in
  stripe-webhook-endpoint.ts; drop the dep + tsconfig project-ref
- enterprise: localize the contract at src/billing/webhook-contract.ts (it owns
  the canonical class), keep `implements`, export the contract, retarget the
  assignability test, drop the dep — keeps the published .d.ts self-contained
- Revert SMI-5066 publish-pipeline registration (publish.yml job + needs/
  predicates + smoke loop + gh-release + summary; version-utils PACKAGE_SPECS;
  verify-publish-deps PACKAGES; 3 release/audit tests)
- Remove billing-types from root tsconfig.json + package-lock.json

Verification (worktree container):
- enterprise tsc --noEmit: 0 errors
- mcp-server: no billing-types resolution errors (pre-existing zod drift only,
  present on origin/main where CI is green per feedback_main_branch_claims_via_ci)
- audit:standards: 0 failed (Check 48 confirms publish.yml parity at 3 paired clauses)
- 112 script tests pass; assignability test passes
- madge: no new cross-package cycles (15 pre-existing intra-package, matches W5 baseline)

After merge, re-running publish.yml publishes mcp-server@0.5.3 + enterprise@0.1.2,
healing the cli@0.6.3 dangling reference.

--no-verify: pre-commit typecheck trips on pre-existing zod ZodTypeDef drift in
license.gate.ts/tool-dispatch.ts/validation.ts (NOT touched by this PR; present
on origin/main; CI fresh-installs zod 3.25.76 from the lockfile and builds green).
