--- packages/billing-types/CHANGELOG.md
@@ -1,12 +1,0 @@
-# Changelog
-
-All notable changes to `@skillsmith/billing-types` are documented here.
-
-## [Unreleased]
-
-## 0.1.0
-
-- **Initial release**: SMI-5044 — types-only contract package created to
-  resolve the `@skillsmith/mcp-server` ↔ `@smith-horn/enterprise` workspace
-  cycle. Exports `StripeWebhookHandler` interface + `StripeWebhookResult`
-  result shape. No runtime code. Public on npmjs.com under Elastic-2.0.
