--- packages/billing-types/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@skillsmith/billing-types",
-  "version": "0.1.0",
-  "description": "Types-only contract shared between @skillsmith/mcp-server and @smith-horn/enterprise for the Stripe webhook surface. No runtime code.",
-  "type": "module",
-  "main": "./dist/src/index.js",
-  "types": "./dist/src/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/src/index.d.ts",
-      "import": "./dist/src/index.js",
-      "default": "./dist/src/index.js"
-    }
-  },
-  "scripts": {
-    "prepublishOnly": "node ../../scripts/lib/forbid-local-publish.mjs && npm run build",
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "test": "vitest run --passWithNoTests",
-    "typecheck": "tsc --noEmit"
-  },
-  "files": [
-    "dist",
-    "CHANGELOG.md",
-    "README.md"
-  ],
-  "license": "Elastic-2.0",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/smith-horn/skillsmith.git",
-    "directory": "packages/billing-types"
-  },
-  "homepage": "https://github.com/smith-horn/skillsmith#readme",
-  "bugs": {
-    "url": "https://github.com/smith-horn/skillsmith/issues"
-  },
-  "keywords": [
-    "skillsmith",
-    "stripe",
-    "billing",
-    "types",
-    "webhook"
-  ],
-  "devDependencies": {
-    "vitest": "4.1.6"
-  },
-  "engines": {
-    "node": ">=22.22.0"
-  }
-}
