+++ packages/mcp-server/package.json
@@ -34,7 +34,6 @@
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.27.1",
-    "@skillsmith/billing-types": "^0.1.0",
     "@skillsmith/core": "^0.8.0",
     "esbuild": "0.27.2",
     "ulid": "3.0.1"
