+++ packages/mcp-server/tsconfig.json
@@ -20,5 +20,5 @@
   },
   "include": ["src/**/*", "tests/**/*"],
   "exclude": ["node_modules", "dist"],
-  "references": [{ "path": "../core" }, { "path": "../billing-types" }]
+  "references": [{ "path": "../core" }]
 }
