+++ scripts/tests/prepare-release.test.ts
@@ -45,15 +45,15 @@ const mockedExecFileSync = vi.mocked(execFileSync)
 const mockedWriteFileSync = vi.mocked(writeFileSync)
 
 describe('PACKAGE_SPECS configuration', () => {
-  it('should define all five packages', () => {
-    // SMI-5066: billing-types added between mcp-server and cli.
+  it('should define all four packages', () => {
+    // SMI-5119: billing-types removed (unpublishable types-only package;
+    // contract localized into @smith-horn/enterprise).
     const names = PACKAGE_SPECS.map((s) => s.shortName)
     expect(names).toContain('core')
     expect(names).toContain('mcp-server')
-    expect(names).toContain('billing-types')
     expect(names).toContain('cli')
     expect(names).toContain('vscode')
-    expect(PACKAGE_SPECS).toHaveLength(5)
+    expect(PACKAGE_SPECS).toHaveLength(4)
   })
 
   it('should point to existing package.json files', () => {
