+++ tsconfig.json
@@ -17,7 +17,6 @@
     "tsBuildInfoFile": ".tsbuildinfo"
   },
   "references": [
-    { "path": "./packages/billing-types" },
     { "path": "./packages/core" },
     { "path": "./packages/mcp-server" },
     { "path": "./packages/cli" }
